fix db config and test connection again
