+++ public/index.html
@@ -12,5 +12,10 @@
       <button id="uploadButton">Upload</button>
       <script src="./js/import_csv.js"></script>
     </div>
+    <div>
+      <a href="db/dbsettings.html">
+          <button>Configure Database</button>
+      </a>
+    </div>
   </body>
 </html>
